backend/requirements-minimal.txt
# Minimal requirements that work with Python 3.8-3.13
# If you encounter installation issues with Python 3.14, use Python 3.11 or 3.12

# Core web framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Data validation (pinned to avoid build issues)
pydantic>=2.0,<3.0

# File handling
python-multipart
PyPDF2

# Environment
python-dotenv

# LLM providers
google-generativeai
openai

# HTTP client for MCP
httpx

# HTTP client for Skywork Router
requests

# Timezone support for skills
pytz

# Testing
pytest
pytest-asyncio
pytest-cov
